Rassemble des définitions d'étapes de test
