--- e2e/cypress/integration/08-gerer-les-informations-des-membres/selectors.js
@@ -1,8 +1,0 @@
-export const LocalSelectors = {
-  modale: {
-    selector: '[data-test=Modal]',
-  },
-  supprimer: {
-    selector: '[data-test=]',
-  },
-};
